Struct deno_core::AsyncRefCell

pub struct AsyncRefCell<T> { /* fields omitted */ }

Implementations

impl<T> AsyncRefCell<T>

pub fn borrow_sync<M: BorrowModeTrait, R: RcLike<AsyncRefCell<T>>>(
cell: R
) -> Option<AsyncBorrowImpl<T, M>>

Borrow the cell’s contents synchronouslym without creating an intermediate future. If the cell has already been borrowed and either the existing or the requested borrow is exclusive, this function returns None.

impl<T: 'static> AsyncRefCell<T>

pub fn new(value: T) -> Self

Create a new AsyncRefCell that encapsulates the specified value. Note that in order to borrow the inner value, the AsyncRefCell needs to be wrapped in an Rc or an RcRef. These can be created either manually, or by using the convenience method AsyncRefCell::new_rc().

pub fn new_rc(value: T) -> Rc<Self>

pub fn as_ptr(&self) -> *mut T

pub fn into_inner(self) -> T

impl<T: Default + 'static> AsyncRefCell<T>

pub fn default_rc() -> Rc<Self>

impl<T> AsyncRefCell<T>

pub fn borrow(self: &Rc<Self>) -> AsyncRefFuture<T>

pub fn borrow_mut(self: &Rc<Self>) -> AsyncMutFuture<T>

pub fn try_borrow(self: &Rc<Self>) -> Option<AsyncRef<T>>

pub fn try_borrow_mut(self: &Rc<Self>) -> Option<AsyncMut<T>>