Struct deno_core::JsRuntime

pub struct JsRuntime { /* fields omitted */ }

A single execution context of JavaScript. Corresponds roughly to the “Web Worker” concept in the DOM. A JsRuntime is a Future that can be used with an event loop (Tokio, async_std). The JsRuntime future completes when there is an error or when all pending ops have completed.

Pending ops are created in JavaScript by calling Deno.core.opAsync(), and in Rust by implementing an async function that takes a serde::Deserialize “control argument” and an optional zero copy buffer, each async Op is tied to a Promise in JavaScript.

Implementations

impl JsRuntime

pub fn new(options: RuntimeOptions) -> Self

Only constructor, configuration is done through options.

pub fn global_context(&mut self) -> Global<Context>

pub fn v8_isolate(&mut self) -> &mut OwnedIsolate

pub fn inspector(&mut self) -> Option<&mut Box<JsRuntimeInspector>>

pub fn handle_scope(&mut self) -> HandleScope<'_>

pub fn sync_ops_cache(&mut self)

Ensures core.js has the latest op-name to op-id mappings

pub fn op_state(&mut self) -> Rc<RefCell<OpState>>

Returns the runtime’s op state, which can be used to maintain ops and access resources between op calls.

pub fn execute(
&mut self,
js_filename: &str,
js_source: &str
) -> Result<(), AnyError>

Executes traditional JavaScript code (traditional = not ES modules)

The execution takes place on the current global context, so it is possible to maintain local JS state and invoke this method multiple times.

AnyError can be downcast to a type that exposes additional information about the V8 exception. By default this type is JsError, however it may be a different type if RuntimeOptions::js_error_create_fn has been set.

pub fn snapshot(&mut self) -> StartupData

Takes a snapshot. The isolate should have been created with will_snapshot set to true.

AnyError can be downcast to a type that exposes additional information about the V8 exception. By default this type is JsError, however it may be a different type if RuntimeOptions::js_error_create_fn has been set.

pub fn register_op<F>(&mut self, name: &str, op_fn: F) -> OpId where
F: Fn(Rc<RefCell<OpState>>, OpPayload<'_, '_, '_>) -> Op + 'static,

Registers an op that can be called from JavaScript.

The op mechanism allows to expose Rust functions to the JS runtime, which can be called using the provided name.

This function provides byte-level bindings. To pass data via JSON, the following functions can be passed as an argument for op_fn:

op_sync()

op_async()

pub fn add_near_heap_limit_callback<C>(&mut self, cb: C) where
C: FnMut(usize, usize) -> usize + 'static,

Registers a callback on the isolate when the memory limits are approached. Use this to prevent V8 from crashing the process when reaching the limit.

Calls the closure with the current heap limit and the initial heap limit. The return value of the closure is set as the new limit.

pub fn remove_near_heap_limit_callback(&mut self, heap_limit: usize)

pub async fn run_event_loop(
&mut self,
wait_for_inspector: bool
) -> Result<(), AnyError>

Runs event loop to completion

This future resolves when:

there are no more pending dynamic imports

there are no more pending ops

there are no more active inspector sessions (only if wait_for_inspector is set to true)

pub fn poll_event_loop(
&mut self,
cx: &mut Context<'_>,
wait_for_inspector: bool
) -> Poll<Result<(), AnyError>>

Runs a single tick of event loop

If wait_for_inspector is set to true event loop will return Poll::Pending if there are active inspector sessions.

impl JsRuntime

pub fn mod_evaluate(&mut self, id: ModuleId) -> Receiver<Result<(), AnyError>>

Evaluates an already instantiated ES module.

Returns a receiver handle that resolves when module promise resolves. Implementors must manually call run_event_loop() to drive module evaluation future.

AnyError can be downcast to a type that exposes additional information about the V8 exception. By default this type is JsError, however it may be a different type if RuntimeOptions::js_error_create_fn has been set.

This function panics if module has not been instantiated.

pub async fn load_module(
&mut self,
specifier: &ModuleSpecifier,
code: Option<String>
) -> Result<ModuleId, AnyError>

Asynchronously load specified module and all of its dependencies

User must call JsRuntime::mod_evaluate with returned ModuleId manually after load is finished.