Struct deno_core::RcRef

pub struct RcRef<T> { /* fields omitted */ }

An RcRef encapsulates a reference counted pointer, just like a regular std::rc::Rc. However, unlike a regular Rc, it can be remapped so that it dereferences to any value that’s reachable through the reference-counted pointer. This is achieved through the associated method, RcRef::map(), similar to how std::cell::Ref::map() works. Example:

struct Stuff {
  foo: u32,
  bar: String,
}

let stuff_rc = Rc::new(Stuff {
  foo: 42,
  bar: "hello".to_owned(),
});

// `foo_rc` and `bar_rc` dereference to different types, however
// they share a reference count.
let foo_rc: RcRef<u32> = RcRef::map(stuff_rc.clone(), |v| &v.foo);
let bar_rc: RcRef<String> = RcRef::map(stuff_rc, |v| &v.bar);

Implementations

impl<T> RcRef<AsyncRefCell<T>>

pub fn borrow(&self) -> AsyncRefFuture<T>

pub fn borrow_mut(&self) -> AsyncMutFuture<T>

pub fn try_borrow(&self) -> Option<AsyncRef<T>>

pub fn try_borrow_mut(&self) -> Option<AsyncMut<T>>

impl<T: 'static> RcRef<T>

pub fn new(value: T) -> Self

pub fn map<S: 'static, R: RcLike<S>, F: FnOnce(&S) -> &T>(
source: R,
map_fn: F
) -> RcRef<T>