Struct deno_core::RuntimeOptions

pub struct RuntimeOptions {
    pub js_error_create_fn: Option<Rc<JsErrorCreateFn>>,
    pub get_error_class_fn: Option<GetErrorClassFn>,
    pub module_loader: Option<Rc<dyn ModuleLoader>>,
    pub extensions: Vec<Extension>,
    pub startup_snapshot: Option<Snapshot>,
    pub will_snapshot: bool,
    pub create_params: Option<CreateParams>,
    pub v8_platform: Option<UniquePtr<Platform>>,
    pub attach_inspector: bool,
}

Fields
js_error_create_fn: Option<Rc<JsErrorCreateFn>>

Allows a callback to be set whenever a V8 exception is made. This allows the caller to wrap the JsError into an error. By default this callback is set to JsError::create().

get_error_class_fn: Option<GetErrorClassFn>

Allows to map error type to a string “class” used to represent error in JavaScript.

module_loader: Option<Rc<dyn ModuleLoader>>

Implementation of ModuleLoader which will be called when V8 requests to load ES modules.

If not provided runtime will error if code being executed tries to load modules.

extensions: Vec<Extension>

JsRuntime extensions, not to be confused with ES modules these are sets of ops and other JS code to be initialized.

startup_snapshot: Option<Snapshot>

V8 snapshot that should be loaded on startup.

Currently can’t be used with will_snapshot.

will_snapshot: bool

Prepare runtime to take snapshot of loaded code.

Currently can’t be used with startup_snapshot.

create_params: Option<CreateParams>

Isolate creation parameters.

v8_platform: Option<UniquePtr<Platform>>

V8 platform instance to use. Used when Deno initializes V8 (which it only does once), otherwise it’s silenty dropped.

attach_inspector: bool

Create a V8 inspector and attach to the runtime.