Trait deno_core::ModuleLoader

pub trait ModuleLoader {
    fn resolve(
        &self, 
        op_state: Rc<RefCell<OpState>>, 
        specifier: &str, 
        referrer: &str, 
        _is_main: bool
    ) -> Result<ModuleSpecifier, AnyError>;

    fn load(
        &self, 
        op_state: Rc<RefCell<OpState>>, 
        module_specifier: &ModuleSpecifier, 
        maybe_referrer: Option<ModuleSpecifier>, 
        is_dyn_import: bool
    ) -> Pin<Box<ModuleSourceFuture>>;

    fn prepare_load(
        &self, 
        _op_state: Rc<RefCell<OpState>>, 
        _load_id: ModuleLoadId, 
        _module_specifier: &ModuleSpecifier, 
        _maybe_referrer: Option<String>, 
        _is_dyn_import: bool
    ) -> Pin<Box<dyn Future<Output = Result<(), AnyError>>>> { ... }
}

Required methods

fn resolve(
&self,
op_state: Rc<RefCell<OpState>>,
specifier: &str,
referrer: &str,
_is_main: bool
) -> Result<ModuleSpecifier, AnyError>

Returns an absolute URL. When implementing an spec-complaint VM, this should be exactly the algorithm described here: https://html.spec.whatwg.org/multipage/webappapis.html#resolve-a-module-specifier

is_main can be used to resolve from current working directory or apply import map for child imports.

fn load(
&self,
op_state: Rc<RefCell<OpState>>,
module_specifier: &ModuleSpecifier,
maybe_referrer: Option<ModuleSpecifier>,
is_dyn_import: bool
) -> Pin<Box<ModuleSourceFuture>>

Given ModuleSpecifier, load its source code.

is_dyn_import can be used to check permissions or deny dynamic imports altogether.

Provided methods

fn prepare_load(
&self,
_op_state: Rc<RefCell<OpState>>,
_load_id: ModuleLoadId,
_module_specifier: &ModuleSpecifier,
_maybe_referrer: Option<String>,
_is_dyn_import: bool
) -> Pin<Box<dyn Future<Output = Result<(), AnyError>>>>

This hook can be used by implementors to do some preparation work before starting loading of modules.

For example implementor might download multiple modules in parallel and transpile them to final JS sources before yielding control back to the runtime.

It’s not required to implement this method.