Trait deno_core::Resource

pub trait Resource: Any + 'static {
    fn name(&self) -> Cow<'_, str> { ... }

    fn close(self: Rc<Self>) { ... }
}

All objects that can be store in the resource table should implement the Resource trait.

Provided methods

fn name(&self) -> Cow<'_, str>

Returns a string representation of the resource which is made available to JavaScript code through op_resources. The default implementation returns the Rust type name, but specific resource types may override this trait method.

fn close(self: Rc<Self>)

Resources may implement the close() trait method if they need to do resource specific clean-ups, such as cancelling pending futures, after a resource has been removed from the resource table.

Loading content...
Implementations

impl dyn Resource

pub fn downcast_rc<'a, T: Resource>(self: &'a Rc<Self>) -> Option<&'a Rc<T>>